bug fixes and LargeDataSyncTest
LargeDataSyncTest is now functional

+++ NLE/NLCore/NL_MSDistributor.inl
@@ -132,10 +132,10 @@ namespace NLE
 			template<typename T>
 			void MSDistributor<T>::localRemove(uint_fast32_t index)
 			{
-				uint_fast32_t size = _data.size();
+				size_t size = _data.size();
 				assert(index < size);
 
-				uint_fast32_t lastOccupied = size - 1;
+				size_t lastOccupied = size - 1;
 				if (index != lastOccupied)
 				{
 					_data[index] = _data[lastOccupied];
